streamline files
