--- include_tilt/tilted_column.h
@@ -1,24 +1,0 @@
-#include <cmath>
-#include <iostream>
-#include <vector>
-#include <numeric>
-#include "Source_functions.h"
-
-struct ijk
-{
-    int i;
-    int j;
-    int k;
-};
-
-inline int sign(Float value)
-{
-    return (Float(0.) < value) - (value < Float(0.));
-}
-
-void tilted_path(std::vector<Float>& xh, std::vector<Float>& yh,
-                 std::vector<Float>& zh, std::vector<Float>& z,
-                 Float sza, Float azi,
-                 Float x_start, Float y_start,
-                 std::vector<ijk>& tilted_path,
-                 std::vector<Float>& dz_tilted);
