+++ src_test/CMakeLists.txt
@@ -43,6 +43,5 @@ target_link_libraries(test_rte_rrtmgp rte_rrtmgp ${LIBS} m)
 
 add_executable(test_rt_tilt_prep 
     test_rt_tilt_prep_mc.cpp
-    ../src_tilt/tilted_column.cpp
-    ../src_tilt/tilt_prep_utils.cpp)
+    ../src_tilt/tilt_utils.cpp)
 target_link_libraries(test_rt_tilt_prep rte_rrtmgp ${LIBS} m)
