+++ src_test/test_rt_tilt_prep_mc.cpp
@@ -18,7 +18,6 @@
 
  #include <boost/algorithm/string.hpp>
  #include <chrono>
- #include <random>
  #include <cmath>
  #include <iomanip>
  #include <cuda_profiler_api.h>
@@ -27,358 +26,9 @@
  #include "Netcdf_interface.h"
  #include "Array.h"
  #include "Gas_concs.h"
- #include "tilted_column.h"
- #include "tilted_column_bulk.h"
- #include "tilt_prep_utils.h"
+ #include "tilt_utils.h"
  #include "types.h"
- #include <omp.h>
- 
- void tica_tilt(
-     const Float sza, const Float azi,
-     const int n_col_x, const int n_col_y, const int n_col,
-     const int n_lay, const int n_lev, const int n_z_in, const int n_zh_in ,
-     Array<Float,1> xh, Array<Float,1> yh, Array<Float,1> zh, Array<Float,1> z,
-     Array<Float,2> p_lay, Array<Float,2> t_lay, Array<Float,2> p_lev, Array<Float,2> t_lev, 
-     Array<Float,2> lwp, Array<Float,2> iwp, Array<Float,2> rel, Array<Float,2> dei, 
-     Gas_concs gas_concs,
-     Array<Float,2>& p_lay_out, Array<Float,2>& t_lay_out, Array<Float,2>& p_lev_out, Array<Float,2>& t_lev_out, 
-     Array<Float,2>& lwp_out, Array<Float,2>& iwp_out, Array<Float,2>& rel_out, Array<Float,2>& dei_out, 
-     Gas_concs& gas_concs_out, 
-     std::vector<std::string> gas_names,
-     bool switch_cloud_optics, bool switch_liq_cloud_optics, bool switch_ice_cloud_optics
- )
- {
-        ////// SETUP FOR CENTER START POINT TILTING //////
-        Array<ijk,1> center_path;
-        Array<Float,1> center_zh_tilt;
-        tilted_path(xh.v(),yh.v(),zh.v(),z.v(),sza,azi, 0.5, 0.5, center_path.v(), center_zh_tilt.v());
-        
-        int n_zh_tilt_center = center_zh_tilt.v().size();
-        int n_z_tilt_center = n_zh_tilt_center - 1;
-    
-        tilt_fields(n_z_in, n_zh_in, n_col_x, n_col_y,
-            n_z_tilt_center, n_zh_tilt_center, n_col,
-            zh, z,
-            center_zh_tilt, center_path,
-            &p_lay_out, &t_lay_out, &p_lev_out, &t_lev_out, 
-            gas_concs_out, gas_names
-        );
-        
-        int idx_hold = 2*(n_z_tilt_center - n_z_in);
-        if ((n_z_tilt_center - idx_hold) % 2 != 0) {
-            idx_hold--;
-        }
-    
-        int compress_lay_start_idx_center = (n_z_tilt_center - idx_hold);
-        if (compress_lay_start_idx_center < 0) {
-            throw std::runtime_error("compress_lay_start_idx is negative - SZA too high.");
-        }
-        compress_fields(compress_lay_start_idx_center, n_col_x, n_col_y,
-                    n_z_in, n_zh_in, n_z_tilt_center,
-                    &p_lay_out, &t_lay_out, &p_lev_out, &t_lev_out, 
-                    gas_concs_out, gas_names);
-        
-        ////// SETUP FOR RANDOM START POINT TILTING //////    
-        if (switch_cloud_optics)
-        {
-            std::vector<std::pair<Float, Float>> x_y_start_arr(n_col);
-            std::vector<Array<ijk,1>> by_col_paths(n_col);
-            std::vector<Array<Float,1>> by_col_zh_tilt(n_col);
-            std::vector<Float> by_col_n_zh_tilt(n_col);
-            std::vector<Float> by_col_n_z_tilt(n_col);
-        
-            std::mt19937_64 rng;
-            uint64_t timeSeed = std::chrono::high_resolution_clock::now().time_since_epoch().count();
-            std::seed_seq ss{uint32_t(timeSeed & 0xffffffff), uint32_t(timeSeed>>32)};
-            rng.seed(ss);
-            std::uniform_real_distribution<double> unif(0.001, 0.999);
-            for (int i = 0; i < n_col; i++)
-            {
-                Float x_point = unif(rng);
-                Float y_point = unif(rng);
-                x_y_start_arr[i] = std::make_pair(x_point, y_point);
-            }
-        
-            Status::print_message("###### Starting Tilting ######");
-            auto time_start = std::chrono::high_resolution_clock::now();
-            #pragma omp parallel for
-            for (int i = 0; i < n_col; i++) {
-                Float x_start = x_y_start_arr[i].first;
-                Float y_start = x_y_start_arr[i].second;
-                Array<ijk,1> path;
-                Array<Float,1> zh_tilt;
-                
-                tilted_path(xh.v(), yh.v(), zh.v(), z.v(), sza, azi, x_start, y_start, path.v(), zh_tilt.v());
-                int n_zh_tilt = zh_tilt.v().size();
-                int n_z_tilt = n_zh_tilt - 1;
-
-                path.set_dims({n_z_tilt}); 
-                zh_tilt.set_dims({n_zh_tilt}); 
-        
-                by_col_paths[i] = path;
-                by_col_zh_tilt[i] = zh_tilt;
-                by_col_n_zh_tilt[i] = n_zh_tilt;
-                by_col_n_z_tilt[i] = n_zh_tilt - 1;
-            }
-            auto time_end = std::chrono::high_resolution_clock::now();
-            auto duration = std::chrono::duration<double, std::milli>(time_end-time_start).count();
-        
-            Status::print_message("Duration tilting columns: " + std::to_string(duration) + " (ms)");
-            Status::print_message("###### Finish Tilting ######");
-        
-            Status::print_message("###### Check Sizes ######");
-            std::vector<Float> by_col_compress_start(n_col);
-            int max_n_z_tilt = 0;
-            for (int i = 0; i < n_col; i++) {
-                const Float n_zh_tilt = by_col_n_zh_tilt[i];
-                const Float n_z_tilt = by_col_n_z_tilt[i];
-                if (n_z_tilt > max_n_z_tilt) {
-                    max_n_z_tilt = n_z_tilt;
-                }
-                int idx_hold = 2*(n_z_tilt - n_z_in);
-        
-                int n_lay_compress = (n_z_tilt - idx_hold) + (idx_hold)/2;
-                int n_lev_compress = n_lay_compress + 1;
-                int compress_lay_start_idx = (n_z_tilt - idx_hold);
-                if (compress_lay_start_idx < 0) {
-                    throw std::runtime_error("compress_lay_start_idx is negative - SZA too high.");
-                }
-                by_col_compress_start[i] = compress_lay_start_idx;
-            }
-    
-            const int total_iterations = n_col_x * n_col_y;
-            std::vector<ColumnResult> col_results(total_iterations);
-        
-            // Do Normalization and Reshaping OUTSIDE OF LOOP
-            Array<Float,2> lwp_norm_reshaped;
-            lwp_norm_reshaped.set_dims({n_z_in, n_col});
-            Array<Float,2> rel_reshaped;
-            rel_reshaped.set_dims({n_z_in, n_col});
-            Array<Float,2> iwp_norm_reshaped;
-            iwp_norm_reshaped.set_dims({n_z_in, n_col});
-            Array<Float,2> dei_reshaped;
-            dei_reshaped.set_dims({n_z_in, n_col});
-    
-            for (int icol = 1; icol <= n_col; ++icol)    
-            {
-                for (int ilay = 1; ilay <= n_zh_in; ++ilay)    
-                {
-                    Float dz = zh({ilay + 1}) - zh({ilay});
-                    if (switch_liq_cloud_optics)
-                    {
-                        (lwp_norm_reshaped)({ilay, icol}) = (lwp)({icol, ilay})/dz;
-                        (rel_reshaped)({ilay, icol}) = (rel)({icol, ilay});
-                    }
-                    if (switch_ice_cloud_optics)
-                    {
-                        (iwp_norm_reshaped)({ilay, icol}) = (iwp)({icol, ilay})/dz;
-                        (dei_reshaped)({ilay, icol}) = (dei)({icol, ilay});
-                    }
-                }
-            }
-            
-            if (switch_liq_cloud_optics){
-                Array<Float,1> lwp_compress;
-                lwp_compress.set_dims({n_z_in});
-                Array<Float,1> rel_compress;
-                rel_compress.set_dims({n_z_in});
-                const std::vector<Float> var_lwp = lwp_norm_reshaped.v();
-                const std::vector<Float> var_rel = rel_reshaped.v();
-    
-                Status::print_message("###### Start Loop ######");
-                auto time_start_loop = std::chrono::high_resolution_clock::now();
-                for (int idx_y = 0; idx_y < n_col_y; idx_y++)
-                {
-                    for (int idx_x = 0; idx_x < n_col_x; idx_x++)
-                    {
-                        int i = idx_x + idx_y * n_col_x;
-                        const Array<ijk,1> tilted_path = by_col_paths[i];
-                        const std::vector<ijk>& tilted_path_v = tilted_path.v();
-                        const Array<Float,1> zh_tilt = by_col_zh_tilt[i];
-                        const int n_z_tilt = by_col_n_z_tilt[i];
-                        const int compress_lay_start_idx = by_col_compress_start[i];
-    
-                        std::vector<Float> var_lwp_tmp(n_z_tilt);
-                        std::vector<Float> var_rel_tmp(n_z_tilt);
-                        std::vector<Float> var_lwp_out(n_z_in);
-                        std::vector<Float> var_rel_out(n_z_in);
-                        
-                        for (int ilay=0; ilay < n_z_tilt; ++ilay)
-                        {
-                            Float dz = zh_tilt({ilay + 1}) - zh_tilt({ilay});
-                            const ijk offset = tilted_path_v[ilay];
-                            const int i_col_new  = ((idx_y+offset.j)%n_col_y) * n_col_x + ((idx_x + offset.i)%n_col_x);
-                            const int idx_in = offset.k + i_col_new*n_z_in;
-                            var_lwp_tmp[ilay] = var_lwp[idx_in] * dz;
-                            var_rel_tmp[ilay] = var_rel[idx_in];
-                        }  
-                        
-                        for (int ilay = 0; ilay < compress_lay_start_idx; ++ilay)
-                        {
-                            var_lwp_out[ilay] = var_lwp_tmp[ilay];
-                            var_rel_out[ilay] = var_rel_tmp[ilay];
-                        }
-                
-                        for (int ilay = compress_lay_start_idx; ilay < n_z_in; ++ilay)
-                        {
-                            const int in_offset = ilay - compress_lay_start_idx;
-                            const int i_lay_in = compress_lay_start_idx + 2 * in_offset;
-                            int num_inputs;
-                            if (ilay < (n_z_in - 1)) {
-                                num_inputs = 2;
-                            } 
-                            else {
-                                num_inputs = ((i_lay_in + 1) == (n_z_tilt - 1)) ? 2 : 3;
-                            }
-                            Float sum = 0.0;
-                            Float t_sum = 0.0;
-                            Float w_sum = 0.0;
-                
-                            for (int k = 0; k < num_inputs; ++k)
-                            {
-                                int in_idx = i_lay_in + k;
-                                sum += var_lwp_tmp[in_idx];
-                                t_sum += var_rel_tmp[in_idx] * var_lwp_tmp[in_idx];
-                                w_sum += var_lwp_tmp[in_idx];
-                            }
-                            var_lwp_out[ilay] = sum;
-                            if (w_sum > 1e-6)
-                            {
-                                var_rel_out[ilay] = t_sum / w_sum;
-                            } 
-                            else 
-                            {
-                                Float avg = 0.0;
-                                for (int k = 0; k < num_inputs; ++k)
-                                {
-                                    int in_idx = (i_lay_in + k);
-                                    avg += var_rel_out[in_idx];
-                                }
-                                var_rel_out[ilay] = avg / num_inputs;
-                            }
-                        }
-    
-                        col_results[i].lwp   = std::move(var_lwp_out);
-                        col_results[i].rel   = std::move(var_rel_out);
-                    }
-                }
-            
-                auto time_end_loop = std::chrono::high_resolution_clock::now();
-                auto duration_loop = std::chrono::duration<double, std::milli>(time_end_loop - time_start_loop).count();
-                Status::print_message("Duration liq loop: " + std::to_string(duration_loop) + " (ms)");
-            }
-    
-            if (switch_ice_cloud_optics){
-                Array<Float,1> iwp_compress;
-                iwp_compress.set_dims({n_z_in});
-                Array<Float,1> dei_compress;
-                dei_compress.set_dims({n_z_in});
-                const std::vector<Float> var_iwp = iwp_norm_reshaped.v();
-                const std::vector<Float> var_dei = dei_reshaped.v();
-            
-                Status::print_message("###### Start Loop ######");
-                auto time_start_loop = std::chrono::high_resolution_clock::now();
-                for (int idx_y = 0; idx_y < n_col_y; idx_y++)
-                {
-                    for (int idx_x = 0; idx_x < n_col_x; idx_x++)
-                    {
-                        int i = idx_x + idx_y * n_col_x;
-                        const Array<ijk,1> tilted_path = by_col_paths[i];
-                        const std::vector<ijk>& tilted_path_v = tilted_path.v();
-                        const Array<Float,1> zh_tilt = by_col_zh_tilt[i];
-                        const int n_z_tilt = by_col_n_z_tilt[i];
-                        const int compress_lay_start_idx = by_col_compress_start[i];
-            
-                        std::vector<Float> var_iwp_tmp(n_z_tilt);
-                        std::vector<Float> var_dei_tmp(n_z_tilt);
-                        std::vector<Float> var_iwp_out(n_z_in);
-                        std::vector<Float> var_dei_out(n_z_in);
-                        
-                        for (int ilay=0; ilay < n_z_tilt; ++ilay)
-                        {
-                            Float dz = zh_tilt({ilay + 1}) - zh_tilt({ilay});
-                            const ijk offset = tilted_path_v[ilay];
-                            const int i_col_new  = ((idx_y+offset.j)%n_col_y) * n_col_x + ((idx_x + offset.i)%n_col_x);
-                            const int idx_in = offset.k + i_col_new*n_z_in;
-                            var_iwp_tmp[ilay] = var_iwp[idx_in] * dz;
-                            var_dei_tmp[ilay] = var_dei[idx_in];
-                        }  
-                        
-                        for (int ilay = 0; ilay < compress_lay_start_idx; ++ilay)
-                        {
-                            var_iwp_out[ilay] = var_iwp_tmp[ilay];
-                            var_dei_out[ilay] = var_dei_tmp[ilay];
-                        }
-                
-                        for (int ilay = compress_lay_start_idx; ilay < n_z_in; ++ilay)
-                        {
-                            const int in_offset = ilay - compress_lay_start_idx;
-                            const int i_lay_in = compress_lay_start_idx + 2 * in_offset;
-                            int num_inputs;
-                            if (ilay < (n_z_in - 1)) {
-                                num_inputs = 2;
-                            } 
-                            else {
-                                num_inputs = ((i_lay_in + 1) == (n_z_tilt - 1)) ? 2 : 3;
-                            }
-                            Float sum = 0.0;
-                            Float t_sum = 0.0;
-                            Float w_sum = 0.0;
-                
-                            for (int k = 0; k < num_inputs; ++k)
-                            {
-                                int in_idx = i_lay_in + k;
-                                sum += var_iwp_tmp[in_idx];
-                                t_sum += var_dei_tmp[in_idx] * var_iwp_tmp[in_idx];
-                                w_sum += var_iwp_tmp[in_idx];
-                            }
-                            var_iwp_out[ilay] = sum;
-                            if (w_sum > 1e-6)
-                            {
-                                var_dei_out[ilay] = t_sum / w_sum;
-                            } 
-                            else 
-                            {
-                                Float avg = 0.0;
-                                for (int k = 0; k < num_inputs; ++k)
-                                {
-                                    int in_idx = (i_lay_in + k);
-                                    avg += var_dei_out[in_idx];
-                                }
-                                var_dei_out[ilay] = avg / num_inputs;
-                            }
-                        }
-            
-                        col_results[i].iwp   = std::move(var_iwp_out);
-                        col_results[i].dei   = std::move(var_dei_out);
-                    }
-                }
-            
-                auto time_end_loop = std::chrono::high_resolution_clock::now();
-                auto duration_loop = std::chrono::duration<double, std::milli>(time_end_loop - time_start_loop).count();
-                Status::print_message("Duration ice loop: " + std::to_string(duration_loop) + " (ms)");
-            }
-            
-            post_process_output(col_results, n_col_x, n_col_y, n_z_in, n_zh_in,
-                &lwp_out, &iwp_out, &rel_out, &dei_out,
-                switch_liq_cloud_optics, switch_ice_cloud_optics);
-    
-        }
-    
-        restore_bkg_profile_bundle(n_col_x, n_col_y, 
-            n_lay, n_lev, n_lay, n_lev, 
-            n_z_in, n_zh_in, n_z_in, n_zh_in,
-            &p_lay_out, &t_lay_out, &p_lev_out, &t_lev_out, 
-            &lwp_out, &iwp_out, &rel_out, &dei_out,
-            gas_concs_out,
-            &p_lay, &t_lay, &p_lev, &t_lev, 
-            &lwp, &iwp, &rel, &dei,
-            gas_concs, 
-            gas_names,
-            switch_liq_cloud_optics, switch_ice_cloud_optics
-        );
- }
- 
+  
  void tilt_input(int argc, char** argv)
  {
     std::vector<std::string> gas_names = {
@@ -537,16 +187,6 @@
         }
     }
     ////// FINISH READING THE ATMOSPHERIC DATA //////
-    // if t lev all 0, interpolate from t lay
-    if (*std::max_element(t_lev.v().begin(), t_lev.v().end()) <= 0) {
-    for (int i = 1; i <= n_col; ++i) {
-        for (int j = 2; j <= n_lay; ++j) {
-            t_lev({i, j}) = (t_lay({i, j}) + t_lay({i, j - 1})) / 2.0;
-        }
-        t_lev({i, n_lev}) = 2 * t_lay({i, n_lay}) - t_lev({i,n_lay});
-        t_lev({i, 1}) = 2 * t_lay({i, 1}) - t_lev({i,2});
-        }
-    }
 
     Array<Float,2> t_lay_out = t_lay;
     Array<Float,2> t_lev_out = t_lev;
